change: data classも分けました
