+++ app/src/main/java/com/example/tsundokun/ui/home/component/data/Screen.kt
@@ -1,0 +1,8 @@
+package com.example.tsundokun.ui.home.component.data
+
+/*
+ * タブの種類
+ */
+enum class Screen {
+    ALL, FAVORITE
+}
