+++ app/src/main/java/com/example/tsundokun/ui/home/component/data/WebPage.kt
@@ -1,0 +1,13 @@
+package com.example.tsundokun.ui.home.component.data
+
+/*
+ * Webページの情報のデータクラス
+ */
+data class WebPage(
+    val title: String?,
+    val ogpImageUrl: String?,
+    val faviconImageUrl: String?,
+    val link: String?,
+    val isFavorite: Boolean,
+    val id: String,
+)
